CSV: Emit all lines on completion
As @kciesielski described in #315 he encountered a bug which sometimes swallowed CSV
lines.
The issue occurs on stream completion when more than on CSV line is not parsed yet and the
last line has no line ending.
I haven't been able to re-create the issue in a proper test case as it is connected to
stream execution timings.