Explicitly expect some logging
Makes explicit that we expect this logging, and suppresses it
when running the tests.

We still see some errors in the logging because sshj also logs
to slf4j itself. Perhaps we want something similar for that
output? Not sure if slf4j also has a kind of testkit/approach for
this?