DynamoDB: Fix parallelism #876
This PR fixes the parallelism in DynamoDB so that it is applied to the HTTP connection pool.

I upped the default parallelism to 32 as this is the nearest power of 2 below 50, which is the default that [AWS used in their own SDK before deprecating the methods](https://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/dynamodbv2/AmazonDynamoDBAsyncClient.html). I've run up to 128 connections in production (by setting the akka-http config propertis) without any problems so 32 seems perfectly safe.