Support streaming of json-values.
This enables parsing a json-structure in a streaming way. It allows to pick the values wanted from the original structure via `JsonPath`, a notation to refer to a specific part of a json-structure.

It never reads the original value in memory entirely (unless that is required by the given json-path), but instead parses the incoming values in chunks. This is especially helpful when consuming APIs, which put the results you're interested in in some nested structure, which isn't simply streamable.

A common example would be:

```
{
  "rows": [...]
}
```

Where the application is only interested in the values contained in `rows`, preferably as they arrive (over a TCP stream for example).