Unix Domain Socket - Improve halfClose handling (#1142)
This PR includes a number of improvements to the `halfClose` handling of the Unix Domain Socket. `halfClose` is particularly relevant for clients (and enabled by default), as they will close their write side to signal end of input, but will still wish to read data from the server.

The main changes here are to call `shutdownOutput` and `shutdownInput` where appropriate, instead of closing the channel. Additionally, a bug in JNR is worked around which was causing some NIO events to be lost.

Fixes https://github.com/akka/alpakka/issues/1139