JMS: Adjust test to fail-fast nature of producer
The previous implementation of the test was flaky, as it did assume that
on failure of one send operation, the producer stage would emit all
in-flight messages before the failure and then fail the stage.

The actual behavior is fail-fast: failure of one send operation causes
immediate failure of the stage. This commit adjusts the test to this
behavior (that corresponds to MapAsync's behavior).