Introduces packet routing (formally the packet id allocator)
I've just realised that there's a ton of code in the Subscriber that isn't required i.e. once subscribed, the actor can die - it doesn't need to stay around. Instead, I need to implement the Consumer FSM to receive packets - which is pretty much most of the Subscriber right now.