Supervision loses the failure field
As per the bug described at https://github.com/akka/akka/issues/25803#issuecomment-430949523, I’ve now removed supervision and will just put up with the logging. We could easily revert this commit at a later stage if required i.e. when the bug is fixed, and if we still deem the logging to be a real problem.