AWS SQS: Source - Allow configuring visibility timeout (#1287)
**Feature**: 
Allow setting the visibility timeout parameter in the SQS `Receive` request.
this means that in each `Receive` request it will override the default visibility timeout from SQS queue configuration. if one will not set it, the default will be taken.

The requirement:
if one knows the maximum time it takes to process a message and want to set it in the application (and not in the SQS queue console) this change will reduce the number of calls. So, instead of 2 calls - `Receive` and `ChangeVisibilityTimeout`, we will just run the `Receive`

We, At the company I'm working for, started a code migration from our old SQS client to Alpakka.
during the migration we found out this issue, which required us to run 2 consecutive calls - receive the message and change the visibility timeout. I think it's an overhead we can reduce.