Introducing tell
Session objects can now be told to perform a command directly. The command at this point is just PUBLISH, and the rationale is that the session object is the only thing that can guarantee the QoS requirements of a PUBLISH.

Prior to this work, PUBLISH commands were communicated with a command flow, which is often associated with a network connection. The problem with that approach was that PUBLISH commands could be lost if a network connection was lost i.e. the command is consumed, sent via TCP, TCP fails, and the command is lost.