Avoid a race condition when creating child actors
We observed an `InvalidActorNameException` when creating child actors. This could have been due to the following sequence when receiving a “Publish Received Locally” message:

1. prl received, producer actor created
2. producer actor terminates and sends a termination message
3. prl received before termination message is received, the parent actor creates another producer
4. the termination message is received and then attempts to create another producer with the same name

The solution is to explicitly track active consumers and producers rather than rely on another data structure such as `context.children`, which will be updated in response to other events.