Corrected misinterpretation of behaviour setup
I had thought that Behaviors.setup would only get called once for an actor. This is not the case - it is called each time a behaviour is transitioned to.

There was only one place where my misunderstanding mattered, which was where a promise could be tried to be completed multiple times.