MQTT streaming: Resilient client publications (#1315)
Client publications wouldn’t previously withstand a connection being lost, whereas server publications would. I’ve now aligned the client publication handling with the server publication handling.

In addition, session objects are now told to perform a publish command directly. The rationale is that the session object is the only thing that can guarantee the QoS requirements of a PUBLISH. I use the "tell" syntax of actors as I feel that it has the same meaning i.e. "communicate directly". For example:

```scala
clientSession ! Command(Publish(topic, ByteString("ohi")))
```

Prior to this work, PUBLISH commands were communicated with a command flow, which is often associated with a network connection. The problem with the approach was that PUBLISH commands could be lost if a network connection was lost i.e. the command is consumed, sent via TCP, TCP fails, and the command is lost. To substantiate this, TCP flows were often coupled with `RestartFlow`, which states, _"The restart process is inherently lossy, ..."_.

I've tested these changes with an external program and observed that publications recover from a loss of network connectivity.