Akka 2.5.19 upgrade
Affects Akka Typed usage in mqtt-streaming given some API changes.

Also found a bug in matt-streaming along the way in that it was incorrect in handling client connection failures. We should handle a client connection actor terminating only when it terminates gracefully or with an exception that we expect. If we get an unexpected exception then we should not try to handle it. We should also notify observers of the client connection being terminated no matter whether it stops gracefully or not.