Support Whitesource groups (#1392)
This introduces the `whitesourceGroup` setting to be able to select a project in Whitesource into which a sbt project reports its dependencies.
By default, all Alpakka modules report into the Community group. When modules are moved to Incubating or Supported, the Supported group has to be selected for that module.
Whitesource is configured for this already.