MQTT streaming: More connection failure hardening (#1394)
We've been testing the MQTT connector in a production environment, hence discovering a few bugs. The production environment consists of a very low powered CPU and so the scenarios being fixed are largely as a result of a completely different timing behaviour to that of much faster machines.

The collection of commits here largely address an oversight on my part where I had misunderstood `flatMapMerge` to complete on a `Source` completing. It doesn't, and it shouldn't, and I don't know what I was thinking. :-)

I've now introduced a kill switch so that when a disconnection occurs within a client or server session then their associated command flows are terminated. I've enhanced a test to prove this out.