Fixup the continuation of a session
Prior to this commit, if a server side session was publishing to a client and the client disconnected, and then subsequently reconnected, its publications would be lost. This is because I was closing over some actor state within a future. Whoops.

I’ve now written a comprehensive test for this scenario, along with some additional subscription tests along the way.