Fail the client when no CONNACK or SUBACK
Prior to this commit, a lack of a CONNACK would result in the client stream completing, but with no explanation. We now fail the client stream with an appropriate exception.

Also, the lack of a SUBACK would be silently ignored. This commit makes the lack of a SUBACK consistent with a lack of a CONNACK in that the client stream will be failed with an appropriate exception.