Refactor S3 API to static factory methods (#1395)
Refactors S3 connector to embrace the reusability of Akka Stream operators. Instead of creating instances of `S3Client` that then are used to create `Source/Flow/Sink`s, all of the factories are moved to the `S3` object. If connector implementation needs an ActorSystem, ExecutionContext or S3Settings, all of those are taken from the materializer.

This also enables `MinioS3IntegrationSpec` and configures `minio` docker image for that test.

Continuation of #1323