Separate producer/consumer timeouts
Producers of PUBLISH should tend to need to re-send messages faster than the consumer may be able to process them, otherwise the consumer may fail the connection entirely and cascade the problem. Sending multiple publications is benign with MQTT and has the positive effect of resetting the consumer’s timers. This works particularly well on machines where the JVM startup is very slow e.g. a single core Atom processor.

As a side note, Paho’s client has the same separation of concerns in terms of retry interval and receive timeout. We also have the same values of 15 seconds and 30 seconds respectively.