Tolerate DUP PUBLISH messages on first connecting
It is quite possible that a client can receive a PUBLISH message immediately on connection and prior to a subscription given the prior establishment of a server-side session. In this case, the PUBLISH message will often have the DUP flag set. Prior to this commit, the DUP flag would cause a packet router failure given that the routing decision was being performed out of step with where the routing state is managed. This commit moves that routing logic to the correct place.

A new test has been added for the use-case described above and existing coverage also validates the DUP processing.