MQTT streaming: Tolerate DUP PUBLISH messages on first connecting (#1436)
It is quite possible that a client can receive a `PUBLISH` message immediately on connection and before a subscription occurs given the prior establishment of a server-side session. In this case, the `PUBLISH` message will often have the `DUP` flag set. Prior to this PR, the `DUP` flag would cause a packet router failure given that the routing decision was being performed out of step with where the routing state is managed. This PR moves that routing logic to the correct place.

A new test has also been added for the use-case described above along with existing coverage validating the `DUP` processing.