Client/server stashing and disconnectivity
For the client state, I know ensure that any stashed messages are carried through while in a disconnected state. In production, I observed a situation where a consumer completed but didn’t deregister from the internal data structure given that the actor was in a disconnected state.

For the server state, we now ensure that stashed messages are dispatched and that the remote source queue is completed - very similar to how the client behaves. I think we had a resource leak here before… Also, we ignore “connection lost” messages when we are in a disconnected state - we know that the connection is lost at this point, so we want to avoid failing in that scenario.

Both the client and server state data structures now look quite consistent with each other, which makes me feel warm and fuzzy inside.

Finally, there were some errors in the server reconnect test. A different connection id is required each time as should be the case when used in anger.