Consumers post unsubscription
An earlier commit killed off consumers post an unsubscription. According to the MQTT spec, this is incorrect as any PUBLISH messages with a QoS issued by a server must be acknowledged. This commit reverts the change and enhances a test to try and prove the condition, although the condition is hard to test.