Introduces the ability to be notified of server-side SubAck completion
It is important for the server to know when it is the right time to start publishing for a newly subscribed topic; too early and the publications will be dropped given that internal state hasn’t yet updated (sometimes!).

This commit introduces the ability for certain commands, presently only server-side SubAck, to complete a process when the SubAck has been dispatched and internal state updated.