mqtt-streaming: Prefer wrapping instead of reissuing packet ids
Packet ids now monotonically increase until they reach the maximum
value, upon which they wrap around to the minimum. Care is taken to
ensure that packet ids that are in use are not issued, as it is
technically possible, albeit rare.

This fixes a number of potential race conditions. For example, consider
a PUBLISH -> PUBACK exchange with QoS1 with the old behavior:

1) Server retransmits an unacknowleged publish
2) Before client receives retransmission, it finally ACKs original
publish.
3) Server receives ACK, and performs new publish, reusing the last id.
4) Client receives retransmission from #1, and ACKs it.
5) Server receives ACK, and misinterprets it as an ack from #3
6) Oh no!

By wrapping around, we've effectively eliminated the changes of this
happening. This is similar to how Linux issues PIDs for processes to
avoid a similar sort of problem.