mqtt-streaming: Fix a race condition in server publishing
The MQTT streaming API allows the server to supply a `Promise` when it
offers a `SUBACK`. The contract here is that when the `Promise` is
completed, it is safe for the server to begin sending `PUBLISH`es as all
of the internal machinery is ready.

However, due to a race condition, `PUBLISH` commands sent in response to
this `Promise` being completed could arrive before the server's
`ClientConnection` has messaged itself the `Subscribed` instance.
If this happens, the `PUBLISH` is ignored.

This introduces another promise that the `ClientConnection`
is responsible for completing when it receives the `Subscribe` message
in response. It then completes this `Promise`, which in turn completes the
user's supplied `Promise`, thus eliminating the race and ensuring that
`PUBLISH`es are not ignored.