MQTT streaming: Fixes wrong Exception thrown when client ping timeouts (#1642)
## Purpose

When mqtt client doesn't send ping request within the expected period (given by the connect keep alive parameter) stream should failed with `ServerConnector.PingFailed` exception but it fails with `ClientConnection.PingFailed` exception. That makes that ActorMqttServerSession command flow doesn't translate the failure to the public api error ActorMqttServerSession.PingFailed (MqttSession.scala::482)

## Background Context

There was an ignored failing unit test due to this bug that however was pointing to a different issue. After this PR this unit test is not failing anymore so I activate it again.