MQTT streaming: SourceQueue backpressure (#1577)
* mqtt streaming - introduce a mechanism to use SourceQueue backpressure

When offering commands to the internal queue, a backpressure strategy is
now used to ensure that elements are not dropped. Instead, we wait for
consumption downstream before continuing.

We want WatchedActorTerminatedException to complete stream processing as before, but not with a failure. A session shutting down isn’t in itself a failure.