Failure to attain QoS is not bad
Prior to this commit, a failure to attain a QoS level i.e. subscribe, was considered bad. The spec treats this condition as one being something for the application to handle e.g. a client can decide to proceed or not given that one of the topics being subscribed was able to be so.

The fix is to deprecate the BadSubAckMessage and always return SubAck. This then permits the Subscriber to process SubAcks whether they are good or bad from the application's perspective.