Automatically unregister packet ids
Prior to this commit, actors associated with packet identifiers were responsible for not only registering with a packet router but also unregistering. The problem with the approach was that actors can sometimes terminate abnormally, and therefore leave some resources hanging around e.g. some uncompleted promises associated with routing.

This commit changes the behaviour so that actors do not have to explicitly unregister with a packet router. By virtue of terminating, an actor's associated packet identifier and other resources can be cleaned up. Not only does this avoid unwanted memory retention, but as a bonus, it simplifies the usage of the packet routers.