MQTT streaming: Treat ConsumeFailed as a problem (#1852)
When we get a ConsumeFailed exception in the Consumer, this is a problem. The problem should cause the client to shutdown in the same way that subscription, or unsubscription should fail. Same for the server.

This fix addresses a long-standing issue we've had on a super slow machine where ConsumeFailed occurs from time to time, and the stream effectively hangs because the calling wasn't made aware of the problem and couldn't restart the connection etc.