Ignore ConsumeFailed
These changes unwind a previous PR where ConsumeFailed should cause the shutdown of a connection. In hindsight, the original view of letting ConsumeFailed drop on the floor is the right approach. What has changed since then is that, if a publish is acknowledged but cannot be routed then its completion promise will fail. This wasn't happening before.