Update test expectations for Splitter
Since Akka 2.6.2 (https://github.com/akka/akka/pull/28467), 'filter' has
a 1-element buffer, to be able to 'throw away' filtered-out messages even
when there is no downstream demand.

This changes the behavior in a way that makes the SplitterSpec fail.

I updated the tests to make looser assumptions about the amount of buffering,
and only look at what happens when one of the filter slots is filled.
Unfortunately/surprisingly the test now doesn't appear to pass on Akka 2.5,
I'm proposing to put it on 'ignore' until we start requiring Akka 2.6.

/cc @tg44