Make DelaySpec complete only after processing all elements
When the input is completed, the 'delayed' side of the split and the 'non-delayed' side of the split are both completed, and whichever completes first will complete the output.

This means that when there are no delayed elements, and the input completes, this completion may overtake elements still on the 'non-delaying' side, as is visible in test failures.

One solution might be to set eagerComplete to 'false'. This will mean that when the input completes, the stage will wait for both branches to complete before completing itself. It's not completely obvious if that's desirable: do we want to wait for delayed elements in this case? If not, perhaps we should adapt the DelayFlow stage accordingly.

Refs #2361