GCS: Make md5Hash and crc32c optional
The md5Hash and crc32c fields are still mandatory when deserializing
the Google StorageObject. An attempt to make them optional was done in
a previous PR [1].

However, when not defined a "spray.json.DeserializationException" will be thrown.

Make them both optional consistently with most of the other
StorageObjectWriteableJson parameters.

[1]: https://github.com/akka/alpakka/pull/2388